requirements.txt
Flask==3.0.3
Flask-Mail==0.10.0
python-dotenv==1.0.1
google-generativeai==0.7.2
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
click==8.1.7
blinker==1.8.2
itsdangerous==2.2.0
